ZOOM: use size_t for scan hit counts and str size
Functions ZOOM_scanset_term and ZOOM_scanset_display_term have changed,
so that occ (hit count) and len (string length) are now of type size_t
pointer rather than int pointer. This is to be able to represent large
hit counts and to also just to use the proper type for string length
(strlen result).

+++ include/yaz/zoom.h
@@ -251,11 +251,11 @@ ZOOM_connection_scan1(ZOOM_connection c, ZOOM_query startterm);
 
 ZOOM_API(const char *)
 ZOOM_scanset_term(ZOOM_scanset scan, size_t pos,
-                  int *occ, int *len);
+                  size_t *occ, size_t *len);
 
 ZOOM_API(const char *)
 ZOOM_scanset_display_term(ZOOM_scanset scan, size_t pos,
-                          int *occ, int *len);
+                          size_t *occ, size_t *len);
 
 ZOOM_API(size_t)
 ZOOM_scanset_size(ZOOM_scanset scan);
